Upgrade SBT and Scala
