+++ project/Resolvers.scala
@@ -1,0 +1,8 @@
+import sbt._
+
+object Resolvers {
+  val dockerTestKitResolvers = Seq(
+    "whisk" at "https://dl.bintray.com/whisk/maven",
+    "softprops" at "http://dl.bintray.com/content/softprops/maven"
+  )
+}
